GEODE-3240: Moving LuceneSerializer to the public package
