+++ geode-lucene/src/main/java/org/apache/geode/cache/lucene/LuceneIndexFactory.java
@@ -14,7 +14,6 @@
  */
 package org.apache.geode.cache.lucene;
 
-import org.apache.geode.cache.lucene.internal.repository.serializer.LuceneSerializer;
 import org.apache.lucene.analysis.Analyzer;
 
 import java.util.Map;
@@ -81,4 +80,5 @@ public interface LuceneIndexFactory {
    * @param regionPath The region to index. The entries added to this region will be indexes.
    */
   void create(String indexName, String regionPath);
+
 }
