+++ geode-lucene/src/main/java/org/apache/geode/cache/lucene/LuceneSerializer.java
@@ -13,15 +13,18 @@
  * the License.
  */
 
-package org.apache.geode.cache.lucene.internal.repository.serializer;
+package org.apache.geode.cache.lucene;
 
 import java.util.Collection;
 
 import org.apache.lucene.document.Document;
 
+import org.apache.geode.annotations.Experimental;
+
 /**
  * An interface for writing the fields of an object into a lucene document
  */
+@Experimental
 public interface LuceneSerializer {
 
   /**
